% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{as.disag.data}
\alias{as.disag.data}
\title{Function to fit the disaggregation model}
\usage{
as.disag.data(polygon_shapefile, covariate_rasters, polygon_data,
  covariate_data, aggregation_pixels, coords, startendindex, mesh)
}
\arguments{
\item{polygon_shapefile}{SpatialPolygonDataFrame containing the response data}

\item{covariate_rasters}{RasterStack of covariates}

\item{polygon_data}{data.frame with two columns: polygon id and response}

\item{covariate_data}{data.frame with cell id, polygon id and covariate columns}

\item{aggregation_pixels}{vector with value of aggregation raster at each pixel}

\item{coords}{coordinates of the covariate data points}

\item{startendindex}{matrix containing the start and end index for each polygon}

\item{mesh}{inla.mesh object to use in the fit}
}
\description{
Function to fit the disaggregation model
}
\examples{
\dontrun{
  as.disag.data(polygon_data, covariate_data, coords, startendindex, mesh)
 }

}
