% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{disaggregation-deprecated}
\alias{disaggregation-deprecated}
\title{Deprecated functions in disaggregation}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
 \item \code{\link{fit_model}}: This function is deprecated, and will
 be removed in the next version of this package.
}
}
