% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractCoordsForMesh}
\alias{extractCoordsForMesh}
\title{Extract coordinates from raster to use constructing the INLA mesh}
\usage{
extractCoordsForMesh(cov_rasters, covariate_data)
}
\arguments{
\item{cov_rasters}{RasterStack of the covariate rasters}

\item{covariate_data}{data.frame with each covariate as a column an and id column}
}
\description{
Extract coordinates from raster to use constructing the INLA mesh
}
\examples{
\dontrun{
  extractCoordsForMesh(cov_rasters, cov_data)
 }

}
