% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{getAmatrix}
\alias{getAmatrix}
\title{Get Amatrix for field}
\usage{
getAmatrix(mesh, coords)
}
\arguments{
\item{mesh}{mesh used in the model fitting}

\item{coords}{coordinates extracted from raster}
}
\value{
An Amatrix object for the field
}
\description{
Get Amatrix for field
}
