% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{getCoords}
\alias{getCoords}
\title{Get coordinates from raster}
\usage{
getCoords(data)
}
\arguments{
\item{data}{disag.data object}
}
\value{
A matrix of the coordinates of the raster
}
\description{
Get coordinates from raster
}
