% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{getCovariateRasters}
\alias{getCovariateRasters}
\title{Get all covariate rasters in folder and stack them, and crop to polygon data}
\usage{
getCovariateRasters(directory, file_pattern = ".tif$", shape)
}
\arguments{
\item{directory}{filepath containing the rasters}

\item{file_pattern}{Pattern the filenames must match}

\item{shape}{A shape object containing response data}
}
\description{
Get all covariate rasters in folder and stack them, and crop to polygon data
}
\examples{
\dontrun{
  getCovariateRasters('/home/rasters', '.tif$', shape)
 }

}
