% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{getPolygonData}
\alias{getPolygonData}
\title{Extract data from a SpatialPolygonsDataFrame into data.frame of correct structure}
\usage{
getPolygonData(shape, id_var = "area_id", response_var = "response",
  sample_size_var = NULL)
}
\arguments{
\item{shape}{A shape object containing response data}

\item{id_var}{Name of column in shape object with the polygon id}

\item{response_var}{Name of column in shape object with the response data}

\item{sample_size_var}{For survey data, name of column in SpatialPolygonDataFrame object (if it exists) with the sample size data}
}
\description{
Extract data from a SpatialPolygonsDataFrame into data.frame of correct structure
}
\examples{
{
 polygons <- list()
 for(i in 1:100) {
   row <- ceiling(i/10)
   col <- ifelse(i \%\% 10 != 0, i \%\% 10, 10)
   xmin = 2*(col - 1); xmax = 2*col; ymin = 2*(row - 1); ymax = 2*row
   polygons[[i]] <- rbind(c(xmin, ymax), c(xmax,ymax), c(xmax, ymin), c(xmin,ymin))
 }

 polys <- do.call(raster::spPolygons, polygons)
 response_df <- data.frame(area_id = 1:100, response = runif(100, min = 0, max = 10))
 spdf <- sp::SpatialPolygonsDataFrame(polys, response_df)

 getPolygonData(spdf, id_var = 'area_id', response_var = 'response')
}


}
