% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{getStartendindex}
\alias{getStartendindex}
\title{Find the range of pixels in each polygon}
\usage{
getStartendindex(covariates, polygon_data, id_var = "area_id")
}
\arguments{
\item{covariates}{data.frame with each covariate as a column an and id column}

\item{polygon_data}{data.frame with polygon id and response data}

\item{id_var}{string with the name of the column containing the polygon id}
}
\description{
Find the range of pixels in each polygon
}
\examples{
{
covs <- data.frame(area_id = c(1, 1, 1, 2, 2, 3, 3, 3, 3), response = c(3, 9, 5, 2, 3, 6, 7, 3, 5))
response <- data.frame(area_id = c(1, 2, 3), response = c(4, 7, 2))
getStartendindex(covs, response, 'area_id')
}


}
