% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{parallelExtract}
\alias{parallelExtract}
\title{Parallel extraction of raster stack by shape file.}
\usage{
parallelExtract(raster, shape, fun = mean, id = "OBJECTID", ...)
}
\arguments{
\item{raster}{A raster brick or stack}

\item{shape}{A shape object}

\item{fun}{The function used to aggregate the pixel data. If NULL, raw pixel data is returned.}

\item{id}{Name of column in shape object to be used to bind an ID column to output.}

\item{...}{Other arguments to raster::extract}
}
\description{
Parallelisation is performed across rasters, not shapes. 
So this function is only useful if you are extracting 
data from many raster layers.
As the overhead for parallel computation in windows is high
it only makes sense to parallelise in this way.
}
\examples{
 \dontrun{
  cds1 <- rbind(c(-180,-20), c(-160,5), c(-60, 0), c(-160,-60), c(-180,-20))
  cds2 <- rbind(c(80,0), c(100,60), c(120,0), c(120,-55), c(80,0))
  cds3 <- rbind(c(30,20), c(30,50), c(-40,40), c(-10,-30), c(50,10))
  polys <- raster::spPolygons(cds1, cds2, cds3)

  response_df <- data.frame(area_id = c('A', 'B', 'C'), response = c(4, 7, 2))

  spdf <- sp::SpatialPolygonsDataFrame(polys, response_df)

  # Create raster stack
  r <- raster::raster(ncol=36, nrow=18)
  r[] <- 1:raster::ncell(r)
  r <- raster::stack(r, r)

  cl <- parallel::makeCluster(2)
  doParallel::registerDoParallel(cl)
  result <- parallelExtract(r, spdf, fun = NULL, id = 'area_id')
  parallel::stopCluster(cl)
  foreach::registerDoSEQ()
 }
}
