% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.disag.data}
\alias{plot.disag.data}
\title{Plot input data for disaggregation}
\usage{
\method{plot}{disag.data}(x, ...)
}
\arguments{
\item{x}{Object to be plotted}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plotting function for class disag.data (the input data for disaggragation)
}
\details{
Produces three plots: polygon response data, covariates and INLA mesh
}
