% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.disag_prediction}
\alias{plot.disag_prediction}
\title{Plot mean and uncertainty predictions from the disaggregation model results}
\usage{
\method{plot}{disag_prediction}(x, ...)
}
\arguments{
\item{x}{Object of class \emph{disag_prediction} to be plotted.}

\item{...}{Further arguments to \emph{plot} function.}
}
\value{
A list of plots of rasters from the prediction: mean prediction, lower CI and upper CI.
}
\description{
Plotting function for class \emph{disag_prediction} (the mean and uncertainty predictions of the disaggregation fitting).
}
\details{
Produces raster plots of the mean prediction, and the lower and upper confidence intervals.
}
