% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.predictions}
\alias{plot.predictions}
\title{Plot predictions from the disaggregation model results}
\usage{
\method{plot}{predictions}(x, ...)
}
\arguments{
\item{x}{Object of class \emph{predictions} to be plotted.}

\item{...}{Further arguments to \emph{plot} function.}
}
\value{
A list of plots of rasters from the prediction: mean prediction, covariate contribution, field contribution (if used) 
and the iid contribution (if used)
}
\description{
Plotting function for class \emph{predictions} (the mean predictions of the disaggragation fitting).
}
\details{
Produces plots of the mean prediction, and the covariate, field  and iid contribution to the linear predictor.
}
