% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.uncertainty}
\alias{plot.uncertainty}
\title{Plot uncertainty}
\usage{
\method{plot}{uncertainty}(x, ...)
}
\arguments{
\item{x}{Object to be plotted}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot uncertainty
}
