% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_covariate_data}
\alias{plot_covariate_data}
\title{Plot covariate data from RasterStack}
\usage{
plot_covariate_data(x)
}
\arguments{
\item{x}{Object to be plotted}
}
\description{
Plot covariate data from RasterStack
}
