% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_disag_model_data}
\alias{plot_disag_model_data}
\title{Convert results of the model ready for plotting}
\usage{
plot_disag_model_data(x)
}
\arguments{
\item{x}{Object of class \emph{disag_model} to be plotted.}
}
\value{
A list that contains:
 \item{posteriors}{A data.frame of posteriors}
 \item{data}{A data.frame of observed and predicted data}
 \item{title}{The title of the observed vs. predicted plot}
}
\description{
Convert results of the model ready for plotting
}
