% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_polygon_data}
\alias{plot_polygon_data}
\title{Plot polygon data from SpatialPolygonDataFrame}
\usage{
plot_polygon_data(x, zcol = "response")
}
\arguments{
\item{x}{Object to be plotted}

\item{zcol}{name of the response variable to be plotted}
}
\description{
Plot polygon data from SpatialPolygonDataFrame
}
