% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fit.result}
\alias{predict.fit.result}
\title{Predict function to get mean and uncertainty from the model result}
\usage{
\method{predict}{fit.result}(object, newdata = NULL,
  predict_iid = FALSE, N = 100, CI = 0.95, ...)
}
\arguments{
\item{object}{fit.result object returned by fit_model function}

\item{newdata}{If NULL, predictions are made using the data in model_output. 
If this is a raster stack or brick, predictions will be made over this data.}

\item{predict_iid}{If TRUE, any polygon iid effect from the model will be used in the prediction. Default FALSE}

\item{N}{number of realisations. Default: 100}

\item{CI}{confidence interval. Default: 0.95}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Predict function to get mean and uncertainty from the model result
}
\examples{
\dontrun{
predict(fit_result)
}

}
