% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Function to collect all data together for disaggregation model}
\usage{
prepare_data(polygon_shapefile, covariate_rasters, id_var = "area_id",
  response_var = "response", mesh.args = NULL)
}
\arguments{
\item{polygon_shapefile}{SpatialPolygonDataFrame containing the response data}

\item{covariate_rasters}{RasterStack of covariates}

\item{id_var}{Name of column in SpatialPolygonDataFrame object with the polygon id}

\item{response_var}{Name of column in SpatialPolygonDataFrame object with the response data}

\item{mesh.args}{list of parameters that control the mesh structure with the same names as used by INLA}
}
\description{
Function to collect all data together for disaggregation model
}
\examples{
\dontrun{
 cds1 <- rbind(c(-120,-20), c(-100,5), c(-60, 0), c(-100,-60), c(-120,-20))
 cds2 <- rbind(c(30,0), c(50,60), c(70,0), c(70,-55), c(30,0))
 cds3 <- rbind(c(10,20), c(10,50), c(-60,40), c(-30,-30), c(30,10))
 polys <- raster::spPolygons(cds1, cds2, cds3)

 response_df <- data.frame(area_id = c('1', '2', '3'), response = c(4, 7, 2))

 spdf <- sp::SpatialPolygonsDataFrame(polys, response_df)

 # Create raster stack
 r <- raster::raster(ncol=36, nrow=18)
 r[] <- 1:raster::ncell(r)
 cov_rasters <- raster::stack(r, r)

 test_data <- prepare_data(polygon_shapefile = spdf, 
                           covariate_rasters = cov_rasters)
} 
                   
}
