% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.disag.data}
\alias{print.disag.data}
\title{Print function for disaggregation input data}
\usage{
\method{print}{disag.data}(x, ...)
}
\arguments{
\item{x}{Object returned from fit_model}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Function that prints the input data from the disaggregation model
}
\details{
Prints the number of polyons and pixels, the number of pixels in the largest and smallest polygons and summaries of the covariates
}
