% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.disag_data}
\alias{print.disag_data}
\title{Print function for disaggregation input data}
\usage{
\method{print}{disag_data}(x, ...)
}
\arguments{
\item{x}{Object returned from prepare_data.}

\item{...}{Further arguments to \emph{print} function.}
}
\description{
Function that prints the input data from the disaggregation model.
}
\details{
Prints the number of polyons and pixels, the number of pixels in the largest and smallest polygons and summaries of the covariates.
}
