% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.disag_model}
\alias{print.disag_model}
\title{Print function for disaggregation fit result.}
\usage{
\method{print}{disag_model}(x, ...)
}
\arguments{
\item{x}{Object returned from disag_model.}

\item{...}{Further arguments to \emph{print} function.}
}
\description{
Function that prints the result of the fit from the disaggregation model.
}
\details{
Prints the negative log likelihood, model parameters and calculates metrics from in-sample performance.
}
