% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.disag_prediction}
\alias{print.disag_prediction}
\title{Print function for disaggregation prediction}
\usage{
\method{print}{disag_prediction}(x, ...)
}
\arguments{
\item{x}{Object returned from predict.disag_model.}

\item{...}{Further arguments to \emph{print} function.}
}
\description{
Function that prints the prediction from the disaggregation model.
}
\details{
Prints the number of polyons and pixels, the number of pixels in the largest and smallest polygons and summaries of the covariates.
}
