% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.fit.result}
\alias{print.fit.result}
\title{Print function for fit result}
\usage{
\method{print}{fit.result}(x, ...)
}
\arguments{
\item{x}{Object returned from fit_model}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print function for fit result
}
