% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.disag.data}
\alias{summary.disag.data}
\title{Summary function for disaggregation input data}
\usage{
\method{summary}{disag.data}(object, ...)
}
\arguments{
\item{object}{Object returned from fit_model}

\item{...}{Further arguments to \emph{summary} function.}
}
\description{
Function that summarizes the input data from the disaggregation model
}
\details{
Prints the number of polyons and pixels, the number of pixels in the largest and smallest polygons and summaries of the covariates
}
