% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.disag_model}
\alias{summary.disag_model}
\title{Summary function for disaggregation fit result}
\usage{
\method{summary}{disag_model}(object, ...)
}
\arguments{
\item{object}{Object returned from disag_model.}

\item{...}{Further arguments to \emph{summary} function.}
}
\value{
A list of the model parameters, negative log likelihood and metrics from in-sample performance.
}
\description{
Function that summarises the result of the fit from the disaggregation model.
}
\details{
Prints the negative log likelihood, model parameters and calculates metrics from in-sample performance.
}
