% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.disag_prediction}
\alias{summary.disag_prediction}
\title{Summary function for disaggregation prediction}
\usage{
\method{summary}{disag_prediction}(object, ...)
}
\arguments{
\item{object}{Object returned from predict.disag_model}

\item{...}{Further arguments to \emph{summary} function.}
}
\value{
A list of the number of polyons, the number of covariates and summaries of the covariates.
}
\description{
Function that summarizes the prediction from the disaggregation model.
}
\details{
Prints the number of polyons and pixels, the number of pixels in the largest and smallest polygons and summaries of the covariates.
}
