% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mesh.R
\name{build_mesh}
\alias{build_mesh}
\title{Build mesh for disaggregaton model}
\usage{
build_mesh(shapes, mesh.args = NULL)
}
\arguments{
\item{shapes}{shapefile covering the region under investigation}

\item{mesh.args}{list of parameters that control the mesh structure with the same names as used by INLA}
}
\description{
Build mesh for disaggregaton model
}
\examples{
{
 cds1 <- rbind(c(-120,-20), c(-100,5), c(-60, 0), c(-100,-60), c(-120,-20))
 cds2 <- rbind(c(30,0), c(50,60), c(70,0), c(70,-55), c(30,0))
 cds3 <- rbind(c(10,20), c(10,50), c(-60,40), c(-30,-30), c(30,10))
 polys <- raster::spPolygons(cds1, cds2, cds3)

 response_df <- data.frame(area_id = c('1', '2', '3'), response = c(4, 7, 2))

 spdf <- sp::SpatialPolygonsDataFrame(polys, response_df)

 my_mesh <- build_mesh(spdf)
}

}
