% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mesh.R
\name{build_mesh}
\alias{build_mesh}
\title{Build mesh for disaggregaton model}
\usage{
build_mesh(shapes, mesh.args = NULL)
}
\arguments{
\item{shapes}{shapefile covering the region under investigation}

\item{mesh.args}{list of parameters that control the mesh structure with the same names as used by INLA}
}
\description{
Build mesh for disaggregaton model
}
\examples{
\dontrun{
 polygons <- list()
 for(i in 1:100) {
   row <- ceiling(i/10)
   col <- ifelse(i \%\% 10 != 0, i \%\% 10, 10)
   xmin = 2*(col - 1); xmax = 2*col; ymin = 2*(row - 1); ymax = 2*row
   polygons[[i]] <- rbind(c(xmin, ymax), c(xmax,ymax), c(xmax, ymin), c(xmin,ymin))
 }

 polys <- do.call(raster::spPolygons, polygons)
 response_df <- data.frame(area_id = 1:100, response = runif(100, min = 0, max = 10))
 spdf <- sp::SpatialPolygonsDataFrame(polys, response_df)

 my_mesh <- build_mesh(spdf)
}

}
