% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mesh.R
\name{build_mesh}
\alias{build_mesh}
\title{Build mesh for disaggregaton model}
\usage{
build_mesh(shapes, mesh.args = NULL)
}
\arguments{
\item{shapes}{shapefile covering the region under investigation}

\item{mesh.args}{list of parameters that control the mesh structure with the same names as used by INLA}
}
\description{
Build mesh for disaggregaton model
}
\examples{
 \dontrun{
  build_mesh(my_shapefile)
 }


}
