% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractCoordsForMesh}
\alias{extractCoordsForMesh}
\title{Extract coordinates from raster to use constructing the INLA mesh}
\usage{
extractCoordsForMesh(cov_rasters, selectIds = NULL)
}
\arguments{
\item{cov_rasters}{RasterStack of the covariate rasters}

\item{selectIds}{numeric vector containing cell ids to retain. Default NULL retains all cell ids in the covariate rasters}
}
\description{
Extract coordinates from raster to use constructing the INLA mesh
}
\examples{
\dontrun{
  extractCoordsForMesh(cov_rasters, selectIds)
 }

}
