% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Function to fit the disaggregation model}
\usage{
fit_model(data)
}
\arguments{
\item{data}{disag.data object returned by prepare_data function that contains all the necessary objects for the model fitting}
}
\description{
Function to fit the disaggregation model
}
\examples{
 \dontrun{
  fit_model(my_data)
 }

}
