% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Function to fit the disaggregation model}
\usage{
fit_model(data, priors = NULL, family = "gaussian", link = "logit",
  its = 10, field = TRUE, iid = TRUE)
}
\arguments{
\item{data}{disag.data object returned by prepare_data function that contains all the necessary objects for the model fitting}

\item{priors}{list of prior values}

\item{family}{likelihood function: gaussian, binomial or poisson}

\item{link}{link function: logit, log or identity}

\item{its}{number of iterations to run the optimisation for}

\item{field}{boolean to flag spatial field on or off}

\item{iid}{boolean to flag iid effect on or off}
}
\description{
Function to fit the disaggregation model
}
\examples{
\dontrun{
 polygons <- list()
 for(i in 1:100) {
  row <- ceiling(i/10)
  col <- ifelse(i \%\% 10 != 0, i \%\% 10, 10)
  xmin = 2*(col - 1); xmax = 2*col; ymin = 2*(row - 1); ymax = 2*row
  polygons[[i]] <- rbind(c(xmin, ymax), c(xmax,ymax), c(xmax, ymin), c(xmin,ymin))
 }

 polys <- do.call(raster::spPolygons, polygons)
 response_df <- data.frame(area_id = 1:100, response = runif(100, min = 0, max = 10))
 spdf <- sp::SpatialPolygonsDataFrame(polys, response_df)

 r <- raster::raster(ncol=20, nrow=20)
 r[] <- sapply(1:raster::ncell(r), function(x) rnorm(1, ifelse(x \%\% 20 != 0, x \%\% 20, 20), 3))
 r2 <- raster::raster(ncol=20, nrow=20)
 r2[] <- sapply(1:raster::ncell(r), function(x) rnorm(1, ceiling(x/10), 3))
 cov_rasters <- raster::stack(r, r2)

 cl <- parallel::makeCluster(2)
 doParallel::registerDoParallel(cl)
 test_data <- prepare_data(polygon_shapefile = spdf, 
                           covariate_rasters = cov_rasters)
 parallel::stopCluster(cl)
 foreach::registerDoSEQ()
                         
 result <- fit_model(test_data, its = 2)
 }

}
