% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.disag.data}
\alias{plot.disag.data}
\title{These are plot methods for input data and predictions for disaggragation}
\usage{
\method{plot}{disag.data}(x, ...)
}
\arguments{
\item{x}{Object to be plotted}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
These are plot methods for input data and predictions for disaggragation
}
