% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.disag.data}
\alias{plot.disag.data}
\title{Plot input data for disaggregation}
\usage{
\method{plot}{disag.data}(x, ...)
}
\arguments{
\item{x}{Object of class \emph{disag.data} to be plotted.}

\item{...}{Further arguments to \emph{plot} function.}
}
\value{
A list of two plots: the polygon plot (ggplot) and covariate plot (spplot)
}
\description{
Plotting function for class \emph{disag.data} (the input data for disaggragation).
}
\details{
Produces three plots: polygon response data, covariate rasters and INLA mesh.
}
