% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.fit.result}
\alias{plot.fit.result}
\title{Plot results of fitted model}
\usage{
\method{plot}{fit.result}(x, ...)
}
\arguments{
\item{x}{Object to be plotted}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plotting function for class fit.result (the result of the disaggragation fitting)
}
\details{
Produces two plots: results of the fixed effects and in-sample observed vs predicted plot
}
