% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.predictions}
\alias{plot.predictions}
\title{Plot predictions from the disaggregation model results}
\usage{
\method{plot}{predictions}(x, ...)
}
\arguments{
\item{x}{Object to be plotted}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plotting function for class predictions (the mean predictions of the disaggragation fitting)
}
\details{
Produces plots of the mean prediction, and the covariate, field  and iid contribution to the linear predictor
}
