% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.predictions}
\alias{plot.predictions}
\title{Plot predictions}
\usage{
\method{plot}{predictions}(x, ...)
}
\arguments{
\item{x}{Object to be plotted}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot predictions
}
