% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.uncertainty}
\alias{plot.uncertainty}
\title{Plot uncertainty predictions from the disaggregation model results}
\usage{
\method{plot}{uncertainty}(x, ...)
}
\arguments{
\item{x}{Object to be plotted}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plotting function for class uncertainty (the uncertainty predictions of the disaggragation fitting)
}
\details{
Produces plots of the realisations, and the upper and lower credible interval rasters
}
