% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.uncertainty}
\alias{plot.uncertainty}
\title{Plot uncertainty predictions from the disaggregation model results}
\usage{
\method{plot}{uncertainty}(x, ...)
}
\arguments{
\item{x}{Object of class \emph{uncertainty} to be plotted.}

\item{...}{Further arguments to \emph{plot} function.}
}
\description{
Plotting function for class \emph{uncertainty} (the uncertainty predictions of the disaggragation fitting).
}
\details{
Produces a plot of the upper and lower credible interval rasters.
}
