% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_polygon_data}
\alias{plot_polygon_data}
\title{Plot polygon data from SpatialPolygonDataFrame}
\usage{
plot_polygon_data(x, names)
}
\arguments{
\item{x}{Object to be plotted}

\item{names}{list of 2 names: polygon id variable and response variable names}
}
\description{
Plot polygon data from SpatialPolygonDataFrame
}
