% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_model}
\alias{predict_model}
\title{Function to predict from the model result}
\usage{
predict_model(data, model_output)
}
\arguments{
\item{data}{disag.data object returned by prepare_data function}

\item{model_output}{fit.result object returned by fit_model function}
}
\description{
Function to predict from the model result
}
\examples{
\dontrun{
predict_model(data, result)
}

}
