% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_model}
\alias{predict_model}
\title{Function to predict mean from the model result}
\usage{
predict_model(model_output, newdata = NULL)
}
\arguments{
\item{model_output}{fit.result object returned by fit_model function}

\item{newdata}{If NULL, predictions are made using the data in model_output. 
If this is a raster stack or brick, predictions will be made over this data.}
}
\description{
Function to predict mean from the model result
}
\examples{
\dontrun{
predict_model(result)
}

}
