% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Function to collect all data together for disaggregation model}
\usage{
prepare_data(polygon_shapefile, covariate_rasters, id_var = "area_id",
  response_var = "response", mesh.args = NULL)
}
\arguments{
\item{polygon_shapefile}{SpatialPolygonDataFrame containing the response data}

\item{covariate_rasters}{RasterStack of covariates}

\item{id_var}{Name of column in SpatialPolygonDataFrame object with the polygon id}

\item{response_var}{Name of column in SpatialPolygonDataFrame object with the response data}

\item{mesh.args}{list of parameters that control the mesh structure with the same names as used by INLA}
}
\description{
Function to collect all data together for disaggregation model
}
\examples{
\dontrun{
 polygons <- list()
 for(i in 1:100) {
  row <- ceiling(i/10)
  col <- ifelse(i \%\% 10 != 0, i \%\% 10, 10)
  xmin = 2*(col - 1); xmax = 2*col; ymin = 2*(row - 1); ymax = 2*row
  polygons[[i]] <- rbind(c(xmin, ymax), c(xmax,ymax), c(xmax, ymin), c(xmin,ymin))
 }

 polys <- do.call(raster::spPolygons, polygons)
 response_df <- data.frame(area_id = 1:100, response = runif(100, min = 0, max = 10))
 spdf <- sp::SpatialPolygonsDataFrame(polys, response_df)

 r <- raster::raster(ncol=20, nrow=20)
 r[] <- sapply(1:raster::ncell(r), function(x) rnorm(1, ifelse(x \%\% 20 != 0, x \%\% 20, 20), 3))
 r2 <- raster::raster(ncol=20, nrow=20)
 r2[] <- sapply(1:raster::ncell(r), function(x) rnorm(1, ceiling(x/10), 3))
 cov_rasters <- raster::stack(r, r2)

 test_data <- prepare_data(polygon_shapefile = spdf, 
                           covariate_rasters = cov_rasters)
} 
                   
}
