% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Function to collect all data together for disaggregation model}
\usage{
prepare_data(polygon_shapefile, covariate_rasters, id_var = "area_id",
  response_var = "response", mesh.args = NULL)
}
\arguments{
\item{polygon_shapefile}{SpatialPolygonDataFrame containing the response data}

\item{covariate_rasters}{RasterStack of covariates}

\item{id_var}{Name of column in SpatialPolygonDataFrame object with the polygon id}

\item{response_var}{Name of column in SpatialPolygonDataFrame object with the response data}

\item{mesh.args}{list of parameters that control the mesh structure with the same names as used by INLA}
}
\description{
Function to collect all data together for disaggregation model
}
\examples{
 \dontrun{
  prepare_data(my_shapefile, cov_raster_stack)
 }

}
