% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.disag.data}
\alias{print.disag.data}
\title{Print method for disag.data}
\usage{
\method{print}{disag.data}(x, ...)
}
\arguments{
\item{x}{Object returned from fit_model}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print method for disag.data
}
