% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.fit.result}
\alias{print.fit.result}
\title{Print function for disaggregation fit result}
\usage{
\method{print}{fit.result}(x, ...)
}
\arguments{
\item{x}{Object returned from fit_model}

\item{...}{Further arguments to \emph{print} function.}
}
\description{
Function that prints the fit result from the disaggregation model
}
\details{
Prints the negative log likelihood, model parameters and calculates metrics from in-sample performance
}
