% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.disag.data}
\alias{summary.disag.data}
\title{Summary function for disag.data}
\usage{
\method{summary}{disag.data}(object, ...)
}
\arguments{
\item{object}{Object returned from fit_model}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary function for disag.data
}
