% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.fit.result}
\alias{summary.fit.result}
\title{Summary function for disaggregation fit result}
\usage{
\method{summary}{fit.result}(object, ...)
}
\arguments{
\item{object}{Object returned from fit_model}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Function that summarizes the fit result from the disaggregation model
}
\details{
Prints the negative log likelihood, model parameters and calculates metrics from in-sample performance
}
